requirements.txt
langchain
langchain-community
langchain-openai
langchain-anthropic
langgraph
python-dotenv
pydantic
openai
requests
beautifulsoup4